+++ CHANGELOG.md
@@ -1,11 +1,11 @@
-#v0.2.0
+# v0.2.0
 
 - Change the home screen.
 - Add the creature screen.
 - Add the spell screen and return data from API.
 - Add the creature screen and return data from API.
 
-#v0.1.0
+# v0.1.0
 
 - Add core mobile app.
 - Add translation.